Moved files from `web` to `web/api`
